Fixed problem with relations containing node IDs with values that can't fit in 32 bit signed integers

+++ hoot-core/src/main/cpp/hoot/core/io/PbfReader.cpp
@@ -187,7 +187,7 @@ double PbfReader::_convertLat(long lat)
   return .000000001 * (_latOffset + (_granularity * lat));
 }
 
-ElementId PbfReader::_convertToElementId(int id, int memberType)
+ElementId PbfReader::_convertToElementId(long id, int memberType)
 {
   ElementType t;
   switch (memberType)
